requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
aiomysql>=0.2.0
